requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosonic==0.18.0
asyncio==3.4.3
attrs==23.2.0
charset-normalizer==3.1.0
discord-protos==0.0.2
discord.py-self @ git+https://github.com/pc111c/discord.py-self.git@d740474dcd3c764b8f95a002a6db52faecaea974
frozenlist==1.4.1
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
multidict==6.0.5
onecache==0.6.0
protobuf==4.25.3
pyfiglet==1.0.2
rgbprint==4.0.2
tzdata==2024.1
tzlocal==5.2
yarl==1.9.4